diary detail

+++ app/src/main/java/com/boredream/lovebook/data/dto/PageResultDto.kt
@@ -1,9 +1,9 @@
 package com.boredream.lovebook.data.dto
 
 data class PageResultDto<T>(
-    val current: Long,
+    val current: Int,
     val size: Long,
     val total: Long,
-    val pages: Long,
+    val pages: Int,
     val records: List<T>,
 )
